backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
python-multipart
pydantic
pydantic-settings
python-dotenv==1.0.1
mysql-connector-python
passlib[bcrypt]==4.0.1